working towards file-backed buffers everywhere
