+++ test/unit/test_durable_buffer.py
@@ -48,7 +48,7 @@ async def test_unreadable_file(event_loop, tempdir):
 async def test_deletes_messages(event_loop, tempdir):
     b = DurableBuffer(tempdir, "test_deletes_messages", event_loop)
     await b.put(b"some data")
-    ident = b.q._queue[0]
+    ident = b.q._queue[0]["ident"]
     assert await b.get() == b"some data"
     filepath = os.path.join(b._message_path, ident)
     assert not os.path.exists(filepath)
